--- a/ocaml/tests/bench/bench_cached_reads.ml
+++ b/ocaml/tests/bench/bench_cached_reads.ml
@@ -7,8 +7,6 @@ let run () =
 let mutex_workload =
   Bechamel_simple_cli.thread_workload ~before:ignore ~after:ignore ~run
 
-let free_threads (stop, t) = Atomic.set stop true ; Array.iter Thread.join t
-
 let benchmarks =
   Test.make_grouped ~name:"Cached reads"
     [Test.make ~name:"Pool_role.is_master" (Staged.stage Pool_role.is_master)]