@edwintorok 2025-02-05
Interesting, this has failed with
```(cd _build/default/ocaml/libs/xapi-stdext/lib/xapi-stdext-threads && ./scheduler_test.exe)
{Scheduler}
  Testing `Scheduler'.
  This run has ID `BJ5VNE7T'.
  
  .F..
  
  ┌──────────────────────────────────────────────────────────────────────────────┐
  │ [FAIL]        generic          1   test_remove_self.                         │
  └──────────────────────────────────────────────────────────────────────────────┘
  ASSERT same event name
  ASSERT same event name
  ASSERT small time
  FAIL small time
  
     Expected: `300'
     Received: `357'
```

But it doesn't fail like that locally, even when run in a loop, and that code wasn't changed in this PR.